Code refactor on CreateAccountButton.
